Redesigned URL build
To increase code reuse I have moved the build_url method from the child classes and into the parent class, 'Client'.

I also desgined a highly pythonic way of contructing the url.

The children classes are repsonible for setting the values of the api parameters that are located within the parent class.